+++ Reports/CHANGES.md
@@ -6,3 +6,4 @@ Changes in 1.1
 - Add Age breakdown in Calculations.php & CalculationsReports.php
 - Fix breakdown results in ReportsCalculations.fnc.php
 - Add "No Value" to select fields in ReportsCalculations.fnc.php
+- Added English help texts in Help_en.php